+++ app/SearchApi/SearchApi.Core.Test/Adapters/Configuration/ProviderProfileOptionsTest.cs
